+++ repository/src/db_diesel/mod.rs
@@ -19,7 +19,7 @@ mod master_list_line_row;
 mod master_list_name_join;
 mod master_list_row;
 mod name;
-mod name_query;
+mod name_row;
 mod name_store_join;
 mod number_row;
 mod remote_sync_buffer;
@@ -63,8 +63,8 @@ pub use master_list_line::*;
 pub use master_list_line_row::MasterListLineRowRepository;
 pub use master_list_name_join::MasterListNameJoinRepository;
 pub use master_list_row::MasterListRowRepository;
-pub use name::NameRepository;
-pub use name_query::*;
+pub use name::*;
+pub use name_row::NameRowRepository;
 pub use name_store_join::NameStoreJoinRepository;
 pub use number_row::NumberRowRepository;
 pub use remote_sync_buffer::*;
