+++ repository/src/tests.rs.orig
@@ -332,7 +332,7 @@ mod repository_test {
         schema::{PricingRow, NumberRowType, RequisitionRowStatus},
         test_db, CentralSyncBufferRepository, InvoiceLineRepository, InvoiceLineRowRepository,
         InvoiceRepository, ItemRowRepository, MasterListLineRepository, MasterListLineRowRepository,
-        MasterListNameJoinRepository, MasterListRowRepository, NameQueryRepository, NameRepository,
+        MasterListNameJoinRepository, MasterListRowRepository, NameRepository, NameRowRepository,
         NumberRowRepository, OutboundShipmentRepository, RequisitionFilter, RequisitionLineFilter,
         RequisitionLineRepository, RequisitionLineRowRepository, RequisitionRepository,
         RequisitionRowRepository, StockLineRepository, StockLineRowRepository, StoreRowRepository,
@@ -346,7 +346,7 @@ mod repository_test {
         },
         test_db, CentralSyncBufferRepository, ChangelogRepository, InvoiceLineRepository,
         InvoiceLineRowRepository, InvoiceRepository, ItemRowRepository, MasterListLineRowRepository,
-        MasterListNameJoinRepository, MasterListRowRepository, NameQueryRepository, NameRepository,
+        MasterListNameJoinRepository, MasterListRowRepository, NameRepository, NameRowRepository,
         NumberRowRepository, OutboundShipmentRepository, RequisitionLineRepository,
         RequisitionRepository, StockLineRepository, StockLineRowRepository, StockTakeRowRepository,
         StoreRowRepository, UserAccountRepository,
@@ -368,7 +368,7 @@ mod repository_test {
         let connection_manager = get_storage_connection_manager(&settings);
         let connection = connection_manager.connection().unwrap();
 
-        let repo = NameRepository::new(&connection);
+        let repo = NameRowRepository::new(&connection);
         let name_1 = data::name_1();
         repo.insert_one(&name_1).await.unwrap();
         let loaded_item = repo.find_one_by_id(name_1.id.as_str()).await.unwrap();
@@ -385,13 +385,13 @@ mod repository_test {
         let connection = connection_manager.connection().unwrap();
 
         // setup
-        let name_repo = NameRepository::new(&connection);
+        let name_repo = NameRowRepository::new(&connection);
         name_repo.insert_one(&data::name_1()).await.unwrap();
         name_repo.insert_one(&data::name_2()).await.unwrap();
         name_repo.insert_one(&data::name_3()).await.unwrap();
         name_repo.insert_one(&data::name_a_umlaut()).await.unwrap();
 
-        let repo = NameQueryRepository::new(&connection);
+        let repo = NameRepository::new(&connection);
         // test filter:
         let result = repo
             .query(
@@ -541,7 +541,7 @@ mod repository_test {
         let connection = connection_manager.connection().unwrap();
 
         // setup
-        NameRepository::new(&connection)
+        NameRowRepository::new(&connection)
             .insert_one(&data::name_1())
             .await
             .unwrap();
@@ -563,7 +563,7 @@ mod repository_test {
         // setup
         let item_repo = ItemRowRepository::new(&connection);
         item_repo.insert_one(&data::item_1()).await.unwrap();
-        let name_repo = NameRepository::new(&connection);
+        let name_repo = NameRowRepository::new(&connection);
         name_repo.insert_one(&data::name_1()).await.unwrap();
         let store_repo = StoreRowRepository::new(&connection);
         store_repo.insert_one(&data::store_1()).await.unwrap();
@@ -589,7 +589,7 @@ mod repository_test {
         // setup
         let item_repo = ItemRowRepository::new(&connection);
         item_repo.insert_one(&data::item_1()).await.unwrap();
-        let name_repo = NameRepository::new(&connection);
+        let name_repo = NameRowRepository::new(&connection);
         name_repo.insert_one(&data::name_1()).await.unwrap();
         let store_repo = StoreRowRepository::new(&connection);
         store_repo.insert_one(&data::store_1()).await.unwrap();
@@ -695,7 +695,7 @@ mod repository_test {
         let connection = connection_manager.connection().unwrap();
 
         // setup
-        let name_repo = NameRepository::new(&connection);
+        let name_repo = NameRowRepository::new(&connection);
         name_repo.insert_one(&data::name_1()).await.unwrap();
         MasterListRowRepository::new(&connection)
             .upsert_one(&data::master_list_1())
@@ -721,7 +721,7 @@ mod repository_test {
         let connection = connection_manager.connection().unwrap();
 
         // setup
-        let name_repo = NameRepository::new(&connection);
+        let name_repo = NameRowRepository::new(&connection);
         name_repo.insert_one(&data::name_1()).await.unwrap();
         let store_repo = StoreRowRepository::new(&connection);
         store_repo.insert_one(&data::store_1()).await.unwrap();
@@ -760,7 +760,7 @@ mod repository_test {
         let item_repo = ItemRowRepository::new(&connection);
         item_repo.insert_one(&data::item_1()).await.unwrap();
         item_repo.insert_one(&data::item_2()).await.unwrap();
-        let name_repo = NameRepository::new(&connection);
+        let name_repo = NameRowRepository::new(&connection);
         name_repo.insert_one(&data::name_1()).await.unwrap();
         let store_repo = StoreRowRepository::new(&connection);
         store_repo.insert_one(&data::store_1()).await.unwrap();
@@ -803,7 +803,7 @@ mod repository_test {
         item_repo.insert_one(&data::item_1()).await.unwrap();
         item_repo.insert_one(&data::item_2()).await.unwrap();
         item_repo.insert_one(&data::item_service_1()).await.unwrap();
-        let name_repo = NameRepository::new(&connection);
+        let name_repo = NameRowRepository::new(&connection);
         name_repo.insert_one(&data::name_1()).await.unwrap();
         let store_repo = StoreRowRepository::new(&connection);
         store_repo.insert_one(&data::store_1()).await.unwrap();
